changed convention -1 is unordered, first position is 0

+++ instat/clsROperator.vb
@@ -27,23 +27,6 @@ Public Class ROperator
     End Sub
 
     Public Overrides Function ToScript(Optional ByRef strScript As String = "", Optional strTemp As String = "") As String
-
-        Dim i As Integer
-        'Alternative method using OrderedIndices.
-        'If MyBase.OrderedIndices.Count > 0 Then
-        'clsParameters(MyBase.OrderedIndices(0)).bIncludeArgumentName = False
-        'If clsParameters(MyBase.OrderedIndices(0)).bIsOperator AndAlso bBrackets Then
-        'strTemp = strTemp & "(" & clsParameters(MyBase.OrderedIndices(0)).ToScript(strScript) & ")"
-        'Else
-        'strTemp = strTemp & clsParameters(MyBase.OrderedIndices(0)).ToScript(strScript)
-        'End If
-        'End If
-        'For i = 1 To MyBase.OrderedIndices.Count - 1
-        'strTemp = strTemp & Chr(32) & strOperation & Chr(32)
-        'clsParameters(MyBase.OrderedIndices(i)).bIncludeArgumentName = False
-        'strTemp = strTemp & clsParameters(MyBase.OrderedIndices(i)).ToScript(strScript)
-        'Next
-
         'Parameters are sorted in the appropriate order and then the script is built.
         SortParameters()
         If clsParameters(0) IsNot Nothing Then
@@ -66,11 +49,11 @@ Public Class ROperator
         Return MyBase.ToScript(strScript, strTemp)
     End Function
 
-    Public Overrides Sub AddParameter(Optional strParameterName As String = "", Optional strParameterValue As String = "", Optional clsRFunctionParameter As RFunction = Nothing, Optional clsROperatorParameter As ROperator = Nothing, Optional bIncludeArgumentName As Boolean = True, Optional clsParam As RParameter = Nothing, Optional iPosition As Integer = 0)
+    Public Overrides Sub AddParameter(Optional strParameterName As String = "", Optional strParameterValue As String = "", Optional clsRFunctionParameter As RFunction = Nothing, Optional clsROperatorParameter As ROperator = Nothing, Optional bIncludeArgumentName As Boolean = True, Optional clsParam As RParameter = Nothing, Optional iPosition As Integer = -1)
         bIncludeArgumentName = False
         MyBase.AddParameter(strParameterName, strParameterValue, clsRFunctionParameter, clsROperatorParameter, bIncludeArgumentName, clsParam, iPosition)
     End Sub
-    Public Overrides Sub AddParameter(clsParam As RParameter, Optional iPosition As Integer = 0)
+    Public Overrides Sub AddParameter(clsParam As RParameter, Optional iPosition As Integer = -1)
         MyBase.AddParameter(clsParam, iPosition)
     End Sub
 
@@ -85,10 +68,12 @@ Public Class ROperator
         Return Nothing
     End Function
 
-    Public Sub RemoveAllAdditionalParameters() 'Needs to be changed if using OrderedIndices method.
-        SortParameters() 'This is used to bring the parameter with position 1 to the front, then clear all the others using range.
-        If clsParameters.Count > 1 Then
+    Public Sub RemoveAllAdditionalParameters() 'Needs to be edited once first things merged...
+        SortParameters() 'This is used to bring the parameter with position 0 to the front if it exists, then clear all the others using range.
+        If clsParameters.Count > 1 AndAlso clsParameters(0).Position = 0 Then
             clsParameters.RemoveRange(1, clsParameters.Count - 1)
+        Else
+            clsParameters.Clear()
         End If
         OnParametersChanged()
     End Sub
